# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Engineering Supervision System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Activities (Schedule items)
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Owner portal
...
  /projects/{projectId}/generate-owner-link:
    post:
      operationId: generateOwnerLink
      tags: [projects]
      summary: Generate owner access link
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateOwnerLinkBody"
      responses:
        "200":
          description: Generated owner link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerLinkResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, engineer, owner]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - fullName
        - email
        - role
        - createdAt
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, engineer, owner]
      required:
        - username
        - password
        - fullName
        - email
        - role
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        password:
          type: ["string", "null"]

    # Projects
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        ownerEntity:
          type: string
        supervisorEntity:
          type: string
        contractor:
          type: string
        startDate:
          type: string
          format: date
        expectedEndDate:
          type: string
          format: date
        actualEndDate:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [active, completed, delayed, suspended]
        overallProgress:
          type: number
        ownerAccessToken:
          type: ["string", "null"]
        ownerAccessPassword:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - location
        - ownerEntity
        - supervisorEntity
        - contractor
        - startDate
        - expectedEndDate
        - status
        - overallProgress
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        ownerEntity:
          type: string
        supervisorEntity:
          type: string
        contractor:
          type: string
        startDate:
          type: string
          format: date
        expectedEndDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, completed, delayed, suspended]
          default: active
      required:
        - name
        - location
        - ownerEntity
        - supervisorEntity
        - contractor
        - startDate
        - expectedEndDate
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        ownerEntity:
          type: ["string", "null"]
        supervisorEntity:
          type: ["string", "null"]
        contractor:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        expectedEndDate:
          type: ["string", "null"]
          format: date
        actualEndDate:
          type: ["string", "null"]
          format: date
        status:
          type: ["string", "null"]
          enum: [active, completed, delayed, suspended, null]
        overallProgress:
          type: ["number", "null"]

    # Activities
...
    GenerateOwnerLinkBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    OwnerLinkResponse:
      type: object
      properties:
        token:
          type: string
        url:
          type: string
      required:
        - token
        - url

    # Dashboard
...