# lib/api-zod/src/generated/types/generateOwnerLinkBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Engineering Supervision System
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateOwnerLinkBody {
  password: string;
}