# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Engineering Supervision System
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyStatus } from "./updateProjectBodyStatus";

export interface UpdateProjectBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  ownerEntity?: string | null;
  /** @nullable */
  supervisorEntity?: string | null;
  /** @nullable */
  contractor?: string | null;
  /** @nullable */
  startDate?: Date | null;
  /** @nullable */
  expectedEndDate?: Date | null;
  /** @nullable */
  actualEndDate?: Date | null;
  /** @nullable */
  status?: UpdateProjectBodyStatus;
  /** @nullable */
  overallProgress?: number | null;
}